Refactor metrics exposing
